moved manager resource into it's own file
